[R1378] allowing to create a document in a vocabulary detail
