+++ src/component/resource/__tests__/CreateResource.test.tsx
@@ -1,16 +1,11 @@
 import * as React from "react";
 import Ajax from "../../../util/Ajax";
-import {mountWithIntl} from "../../../__tests__/environment/Environment";
 import {intlFunctions} from "../../../__tests__/environment/IntlUtil";
 import Routing from "../../../util/Routing";
 import Routes from "../../../util/Routes";
-import Resource from "../../../model/Resource";
 import {CreateResource} from "../CreateResource";
-import {CreateResourceMetadata} from "../CreateResourceMetadata";
-import {shallow} from "enzyme";
-import TermItFile from "../../../model/File";
-import AppNotification from "../../../model/AppNotification";
-import {Files} from "../document/Files";
+import {mountWithIntl} from "../../../__tests__/environment/Environment";
+import {CreateResourceForm} from "../CreateResourceForm";
 
 jest.mock("../../../util/Routing");
 jest.mock("../../../util/Ajax", () => {
@@ -24,91 +19,22 @@ jest.mock("../../../util/Ajax", () => {
 describe("CreateResource", () => {
     const iri = "http://onto.fel.cvut.cz/ontologies/termit/resource/test";
 
-    let createResource: (resource: Resource) => Promise<string>;
-    let createFile: (file: TermItFile, documentIri : string) => Promise<any>;
-    let uploadFileContent: (fileIri : string, file: File) => Promise<any>;
-    let publishNotification: (notification: AppNotification) => void;
-
     beforeEach(() => {
         Ajax.post = jest.fn().mockImplementation(() => Promise.resolve({ data : iri }));
-        createResource = jest.fn().mockImplementation( () => Promise.resolve( iri ) );
-        createFile = jest.fn();
-        uploadFileContent = jest.fn();
-        publishNotification = jest.fn().mockImplementation( () => Promise.resolve() );
     });
 
     it("returns to Resource Management on cancel", () => {
-        shallow<CreateResource>(<CreateResource createResource={createResource}
-                                                createFile={createFile}
-                                                uploadFileContent={uploadFileContent}
-                                                publishNotification={publishNotification} {...intlFunctions()}/>);
-        CreateResource.onCancel();
+        const wrapper = mountWithIntl(<CreateResource {...intlFunctions()}/>);
+        wrapper.find("#create-resource-cancel").at(0).simulate("click");
         expect(Routing.transitionTo).toHaveBeenCalledWith(Routes.resources);
     });
 
-    it("onCreate is passed data with selected resource type", () => {
-        const wrapper = mountWithIntl(<CreateResource createResource={createResource}
-                                                      createFile={createFile}
-                                                      uploadFileContent={uploadFileContent}
-                                                      publishNotification={publishNotification} {...intlFunctions()}/>);
-        const typeSelectButtons = wrapper.find("button.create-resource-type-select");
-        expect(typeSelectButtons.length).toEqual(2);
-        typeSelectButtons.at(1).simulate("click");
-        const labelInput = wrapper.find("input[name=\"create-resource-label\"]");
-        (labelInput.getDOMNode() as HTMLInputElement).value = "Metropolitan Plan";
-        labelInput.simulate("change", labelInput);
-        const iriInput = wrapper.find("input[name=\"create-resource-iri\"]");
-        (iriInput.getDOMNode() as HTMLInputElement).value = iri;
-        iriInput.simulate("change", iriInput);
-        wrapper.find("button#create-resource-submit").simulate("click");
-        expect(createResource).toHaveBeenCalled();
-        const resource = (createResource as jest.Mock).mock.calls[0][0];
-        expect(resource.types).toBeDefined();
-    });
-
-    it("renders Files component when Document type is selected", () => {
-        const wrapper = mountWithIntl(<CreateResource createResource={createResource}
-                                                      createFile={createFile}
-                                                      uploadFileContent={uploadFileContent}
-                                                      publishNotification={publishNotification}  {...intlFunctions()}/>);
-        wrapper.find("button#create-resource-type-document").simulate("click");
-        expect(wrapper.find(Files).exists()).toBeTruthy();
-    });
-
-    it("renders CreateResourceMetadata component for all resource types except File", () => {
-        const wrapper = mountWithIntl(<CreateResource createResource={createResource}
-                                                      createFile={createFile}
-                                                      uploadFileContent={uploadFileContent}
-                                                      publishNotification={publishNotification} {...intlFunctions()}/>);
-        expect(wrapper.find(CreateResourceMetadata).exists()).toBeTruthy();
-        wrapper.find("button#create-resource-type-document").simulate("click");
-        expect(wrapper.find(CreateResourceMetadata).exists()).toBeTruthy();
-    });
-
-    it("transitions to Resource detail on successful creation", () => {
-        const wrapper = shallow<CreateResource>(<CreateResource
-            createResource={createResource}
-            createFile={createFile}
-            uploadFileContent={uploadFileContent}
-            publishNotification={publishNotification} {...intlFunctions()}/>);
-        wrapper.instance().onCreate(new Resource({iri, label: "Test"}));
-        return Promise.resolve().then(() => {
-            expect(Routing.transitionTo).toHaveBeenCalledWith(Routes.resourceSummary, {
-                params: new Map([["name", "test"]]),
-                query: new Map()
-            });
-        })
-    });
-
-    it("returns resolved identifier on successful creation", () => {
-        const wrapper = shallow<CreateResource>(<CreateResource
-            createResource={createResource}
-            createFile={createFile}
-            uploadFileContent={uploadFileContent}
-            publishNotification={publishNotification} {...intlFunctions()}/>);
-        return wrapper.instance().onCreate(new Resource({iri, label: "Test"})).then(result => {
-            expect(result).toBeDefined();
-            expect(result).toEqual(iri);
-        })
+    it("returns to Resource Management on submit", () => {
+        const wrapper = mountWithIntl(<CreateResource {...intlFunctions()}/>);
+        (wrapper.find(CreateResourceForm).instance().props as any).onSuccess(iri);
+        expect(Routing.transitionTo).toHaveBeenCalledWith(Routes.resourceSummary,  {
+            params: new Map([["name", "test"]]),
+            query: new Map()
+        });
     });
 });
