+++ src/i18n/en.ts
@@ -195,6 +195,7 @@ const en = {
         "vocabulary.document.label": "Document for {vocabulary}",
         "vocabulary.document.attach": "Attach document",
         "vocabulary.document.create": "Create a new document",
+        "vocabulary.document.create.title": "Create a document",
         "vocabulary.document.select": "Select an existing document",
         "vocabulary.document.select.title": "Select a document",
         "vocabulary.document.set": "Set document",
